chore : price 검증 item -> price로 이동 및 price request type int -> BigDecimal 변경

+++ item-api/src/main/java/com/example/itemapi/application/service/ItemService.java
@@ -1,5 +1,7 @@
 package com.example.itemapi.application.service;
 
+import java.math.BigDecimal;
+
 import com.example.itemapi.domain.manager.ItemManager;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -14,7 +16,7 @@ public class ItemService {
      * - 제품을 추가한다.
      * - 같은 제품명이 있는지 확인한다.
      */
-    public void addItem(String name, int price, int stock) {
+    public void addItem(String name, BigDecimal price, int stock) {
         itemManager.addItem(name, price, stock);
     }
 }
